Removing obsolete srvshard.go, moving its stuff to topoproto.
